refactor: Move agent to internal/agent
